#26: o.l.c.m.HasLines matcher
