--- src/test/java/org/llorllale/cactoos/matchers/text/HasLinesTest.java
@@ -1,57 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) for portions of project cactoos-matchers are held by
- * Yegor Bugayenko, 2017-2018, as part of project cactoos.
- * All other copyright for project cactoos-matchers are held by
- * George Aristy, 2018.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package org.llorllale.cactoos.matchers.text;
-
-import org.hamcrest.Description;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.core.IsEqual;
-import org.junit.Test;
-
-/**
- * Test case for {@link HasLines}.
- *
- * @since 0.12
- * @checkstyle JavadocMethodCheck (500 lines)
- */
-@SuppressWarnings("PMD.AvoidDuplicateLiterals")
-public final class HasLinesTest {
-
-    @Test
-    public void matches() {
-        MatcherAssert.assertThat(
-            new HasLines(
-                "A", "C"
-            ).matchesSafely(
-                String.format("A%nB%nC%n"),
-                new Description.NullDescription()
-            ),
-            new IsEqual<>(true)
-        );
-    }
-
-}
